% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbK.R
\name{cbK}
\alias{cbK}
\title{Cost-based Ripley's K and L functions}
\usage{
cbK(loc, cs, directions = 8)
}
\arguments{
\item{loc}{SpatialPoins* object with locations}

\item{cs}{raster object representing the cost surface in
variable \code{band1}.}

\item{directions}{Number of directions in which cells are connected: 4
(rook's case), 8 (queen's case), 16 (knight or king's case) or other (see
\code{\link[raster]{adjacent}})}
}
\value{
A list with elements:
  \itemize{
    \item \code{dm} distance matrix
    \item \code{mds} the MDS fitted object
    \item \code{loc} the MDS locations as a point pattern
    \item \code{K} Ripley's K function
    \item \code{env_K} a Monte Carlo envelope for the K estimate
    \item \code{L} the L function
    \item \code{env_L} a Monte Carlo envelope for the L function
    \item \code{eucl} a list with the same objects for the Euclidean distance
  }
}
\description{
Cost-based Ripley's K and L functions
}
\examples{
  phase1 <- cbK(
    loc = roman_settlements
    cs  = tortosa_cs,
    directions = 4
  )
}

